------------------------------------------------------------------------ r3279 | coomi01 | 2010-07-13 10:30:13 -0400 (Tue, 13 Jul 2010) | 19 lines
UTF8 histogram values are binary elements, and may
contain embeded control characters.

The change adapts an earlier move to allow hex strings
to appear as part of the statistics.

On the output side in opqutils, we mark hex values with
the tag "value HEX:", the important part being the whole
tag appears before the colon, the actual value after.

The input side in opqoptdb is where this change does
it work. In scanning the input buffer, the existing
code moves the point upto the 'v', and then looks for
the colon. We too do just that, but we also then look
distinguish between 'value:' and 'value HEX:'. The code
then calls standard routines to invert the hex string
back to the original key.

------------------------------------------------------------------------
r3280 | zhayu01 | 2010-07-13 12:45:09 -0400 (Tue, 13 Jul 2010) | 6 lines

	Cross integration of change 505966.

	Add new messages for HASHTABLE.

------------------------------------------------------------------------