------------------------------------------------------------------------ r3286 | kiria01 | 2010-07-14 13:00:25 -0400 (Wed, 14 Jul 2010) | 18 lines
Reduce stack recursion for pst_treedup and pst_descinput

Rewrite of pst_treedup and pst_descinput_walk to remove recursion.

The model followed for pst_treedup is that of in-place, depth first replication of the
parse tree using the PST_STK mechanism. The equivalent functions are supported as
before and the comments expanded to describe those not covered.

The model followed for pst_descinput_walk is also depth first so that the tree can
be navigated in reverse when nodes of interest are found. This approach allows for a
more obvious datatype analysis than in the recursive form.

Also initialised the pss_stk_freelist in the parser psq_cbinit
and psq_cbreturn to remove possibility to access freed blocks
from prior memory streams causing random access errors.

------------------------------------------------------------------------
r3287 | thaju02 | 2010-07-14 15:15:19 -0400 (Wed, 14 Jul 2010) | 11 lines

Cross integration of change 505763

Bug 123982: Regression of bug 117775. Bug 117775 reproduces
with the introduction of fix for bug 123748.

This change backs out ingres2006r2 change 504704 and attempts
to readdress bug 123748. This change introduces code to
specifically handle rollback of BMINI/EMINI.

------------------------------------------------------------------------