------------------------------------------------------------------------ r3290 | hanje04 | 2010-07-15 05:45:11 -0400 (Thu, 15 Jul 2010) | 7 lines
Prevent uninstall_ingres.sh from treating Ingres and Vectorwise installations
as the same thing and attempting to remove both as a single entity as it's
default action.

Also fix up "remove all" code as it's was badly broken.

------------------------------------------------------------------------
r3291 | hanje04 | 2010-07-15 11:15:13 -0400 (Thu, 15 Jul 2010) | 4 lines

Change the default for "Date Type Alias" in the modify to add DBMS dialog to
match the value used in the response file.

------------------------------------------------------------------------