------------------------------------------------------------------------ r3303 | horda03 | 2010-07-19 12:45:27 -0400 (Mon, 19 Jul 2010) | 38 lines
Review Board IP  http://reviewboard.ingres.prv/r/4874

Cross integration of change 506064

With CACHE_DYNAMIC enabled, PREPARE'd statements that
need to be recached may cause the FE to see an
E_SC0206 error. No errors are logged in the errlog.

If the PREPARE'd statement contains a scaler sub-select,
then this may involve the use of a Derived table. The
derived table information is stored in the parse tree
as a RNGVAR, but has no (permanent table details). When
the statement is re-cached, the prototype range table
entries are updated, this involves asking RDF to get
the table's modified date - this is only applicable to
TABLES.

pstdsql.c::pst_execute()

If the cache for the session needs to be refreshed,
only call pst_showtab for PST_TABLE rngvar's.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3304 | loera01 | 2010-07-19 16:30:23 -0400 (Mon, 19 Jul 2010) | 47 lines

The ParseConvert() function converts ODBC escape sequence into Ingres or
OpenSQL syntax.  In all cases in which extract() was requested, I found
that the original ODBC syntax could be retained with few modifications.
For example, YEAR(dateTime) was sent to the DBMS as YEAR(dateTime) instead
of extract(YEAR FROM dateTime).  Other cases were considerably more complex
in the ISO form.

Sensitivity to Vectorwise was required in the case of the DAYNAME() scalar,
which returns the name of the day of the week for the date/time value.  The
Ingres dow() scalar nicely for this requirement, but I could find no
comparable equivalent, except for perhaps converting Zeller's Congruence
into a 7-expression case query, which would have consumed over 3,000
characters.  The ODBC returns an error in this case.

ODBC configuration items, such as "Send date/time as Ingres Date" and
"Return empty string DATE values as Null", can be supported for tables with
an Ingres structure, but not a Vectorwise structure.  The ODBC ignores
these attributes when specified, on the assumption that all tables in a
Vectorwise database will have a Vectorwise structure.  This is one instance
of ambiguity between Ingres and Vectorwise capabilities that are far beyond
the scope of this change.

If "Send date/time as Ingres Date" is specified for an Ingres database, the
ODBC reverts to the Ingres form of the date/time scalars.  The same is true
for databases earlier than Ingres 9.1.  The one exception is
TIMESTAMPDIFF(), which reverts to Ingres syntax for 9.1 databases, because
the ISO implementation employs the round() scalar, which is not supported
in Ingres 9.1.

Date/time scalars, which are part of a searched query (i.e., have a column
as an operand), are currently not supported for EA and EDBC databases.  I
am informed by Chris Clark that EA can, in fact, support the Ingres syntax
of these scalars, but that will be addressed in a separate change.

The ODBC follows the same code path for EDBC servers, but EDBC servers are
far less capable of supporting Ingres date/time scalars because date/time
types are limited, especially for timestamps.  Research will be required to
determine what scalars can be supported in EDBc, and what the appropriate
implementation should be.

I threw in sensitivity to Vectorwise for db procedures.  Vectorwise
currently does not support positional parameters, so I made an exception for
Vectorwise.  This allowed three Quiktest tests to pass that failed previously.

The ODBC version was incremented to 3.50.1000.0121.

------------------------------------------------------------------------