------------------------------------------------------------------------ r3304 | loera01 | 2010-07-19 16:30:23 -0400 (Mon, 19 Jul 2010) | 47 lines
The ParseConvert() function converts ODBC escape sequence into Ingres or
OpenSQL syntax.  In all cases in which extract() was requested, I found
that the original ODBC syntax could be retained with few modifications.
For example, YEAR(dateTime) was sent to the DBMS as YEAR(dateTime) instead
of extract(YEAR FROM dateTime).  Other cases were considerably more complex
in the ISO form.

Sensitivity to Vectorwise was required in the case of the DAYNAME() scalar,
which returns the name of the day of the week for the date/time value.  The
Ingres dow() scalar nicely for this requirement, but I could find no
comparable equivalent, except for perhaps converting Zeller's Congruence
into a 7-expression case query, which would have consumed over 3,000
characters.  The ODBC returns an error in this case.

ODBC configuration items, such as "Send date/time as Ingres Date" and
"Return empty string DATE values as Null", can be supported for tables with
an Ingres structure, but not a Vectorwise structure.  The ODBC ignores
these attributes when specified, on the assumption that all tables in a
Vectorwise database will have a Vectorwise structure.  This is one instance
of ambiguity between Ingres and Vectorwise capabilities that are far beyond
the scope of this change.

If "Send date/time as Ingres Date" is specified for an Ingres database, the
ODBC reverts to the Ingres form of the date/time scalars.  The same is true
for databases earlier than Ingres 9.1.  The one exception is
TIMESTAMPDIFF(), which reverts to Ingres syntax for 9.1 databases, because
the ISO implementation employs the round() scalar, which is not supported
in Ingres 9.1.

Date/time scalars, which are part of a searched query (i.e., have a column
as an operand), are currently not supported for EA and EDBC databases.  I
am informed by Chris Clark that EA can, in fact, support the Ingres syntax
of these scalars, but that will be addressed in a separate change.

The ODBC follows the same code path for EDBC servers, but EDBC servers are
far less capable of supporting Ingres date/time scalars because date/time
types are limited, especially for timestamps.  Research will be required to
determine what scalars can be supported in EDBc, and what the appropriate
implementation should be.

I threw in sensitivity to Vectorwise for db procedures.  Vectorwise
currently does not support positional parameters, so I made an exception for
Vectorwise.  This allowed three Quiktest tests to pass that failed previously.

The ODBC version was incremented to 3.50.1000.0121.

------------------------------------------------------------------------
r3305 | hanal04 | 2010-07-20 04:45:28 -0400 (Tue, 20 Jul 2010) | 33 lines

Cross integration of change 506099

Review Board IP  http://reviewboard.ingres.prv/r/4855

Unicode coercion of embedded data is using afe_cvinto()
whenever a char to uni or uni to char coercion is
required. However, Embedded handling of strings is not the
same as that found in the DBMS. Most importantly for this
issue NCHRs need to be null terminated. The existing ADF
routines adu_nvchr_chartouni() and adu_nvchr_unitochar()
do not apply Embedded handling to the results.
This change adds adu_nvchr_embchartouni() and
adu_nvchr_embunitochar() for Embedded handling of unicoercion
cases. The routines are added to ADU to due to header
dependencies. However these routines are STRICTLY for the use
of Embedded data handling.
The new routines terminate CHR and NCHR embedded values.
If we are using a platform where wchar_t is 4 bytes the
UCS2 array of characters returned by the underlying ADF
calls is converted into a series of 4 byte wchar_t characters.
If the output parameter is not big enough the result is truncated.
This is consistent with embedded varchar handling.
adh_dounicoerce() has been updated to return one of three states.
No unicode coercion required, char to uni required or unit to
char required. This allows the correct coercion routine to be
identified and called.
Locally declared nvarchar structures in the ESQLC program
may return incorrect results if the compiler has padded the
structure. To resolve this problem the embedded program must
use "#pragma pack(1)" to ensure the embedded nvarchar
structure is packed.

------------------------------------------------------------------------