------------------------------------------------------------------------ r3305 | hanal04 | 2010-07-20 04:45:28 -0400 (Tue, 20 Jul 2010) | 33 lines
Cross integration of change 506099

Review Board IP  http://reviewboard.ingres.prv/r/4855

Unicode coercion of embedded data is using afe_cvinto()
whenever a char to uni or uni to char coercion is
required. However, Embedded handling of strings is not the
same as that found in the DBMS. Most importantly for this
issue NCHRs need to be null terminated. The existing ADF
routines adu_nvchr_chartouni() and adu_nvchr_unitochar()
do not apply Embedded handling to the results.
This change adds adu_nvchr_embchartouni() and
adu_nvchr_embunitochar() for Embedded handling of unicoercion
cases. The routines are added to ADU to due to header
dependencies. However these routines are STRICTLY for the use
of Embedded data handling.
The new routines terminate CHR and NCHR embedded values.
If we are using a platform where wchar_t is 4 bytes the
UCS2 array of characters returned by the underlying ADF
calls is converted into a series of 4 byte wchar_t characters.
If the output parameter is not big enough the result is truncated.
This is consistent with embedded varchar handling.
adh_dounicoerce() has been updated to return one of three states.
No unicode coercion required, char to uni required or unit to
char required. This allows the correct coercion routine to be
identified and called.
Locally declared nvarchar structures in the ESQLC program
may return incorrect results if the compiler has padded the
structure. To resolve this problem the embedded program must
use "#pragma pack(1)" to ensure the embedded nvarchar
structure is packed.

------------------------------------------------------------------------
r3306 | hanje04 | 2010-07-20 07:00:17 -0400 (Tue, 20 Jul 2010) | 30 lines

Make GUI installer save off response file after successfull installation and
re-use when adding packages to an existing instance

GUI
     - Save response file after successful installation, report if saving fails
     - Add gen_rf_save_name() to generate the full path for saving the
       response file after a successfull installation or finding when
       performing a modify.
     - For modifies, look for saved response file and append new config
       to it instead of generating a new one. Also include any new packages
       being installed.
     - Fix up code to add "misc_ops" so that only options relevent to what
       is actually being installed are written to the response file.
     - Replace popup_error_box() with popup_message_box() and add macros
       for popup_error_box() and popup_warning_box()
     - Added defns. for location to save the response file

RFAPI
     - Add II_RF_OP_APPEND for appending to existing response file rather
       than creating a new one.
     - Make parameter writing more intelligent and check for values
       in each "section" before writing the header, excluding the lot
       if none are set.
     - Replace individual option arrays with rfapiopsinfo[] which
       contains everything except camdb_ops[] which has been kept
       separate as it's a special, rarely used case.
     - Add header for use when appending to existing response file.
       Add structure and typedef for rfapiopsinfo[] array.

------------------------------------------------------------------------