------------------------------------------------------------------------ r3310 | loera01 | 2010-07-20 11:45:11 -0400 (Tue, 20 Jul 2010) | 18 lines
The ODBC parses all queries, looking for keywords that characterize the type
of the query.  The ParseToken() funtion retrieves tokens.  I modified
ParseToken() to recognized comment delimiters and to keep looping through the
query until the first non-comment token is found.  To do this, I broke out the
part of ParseToken() that searches thorugh the token table into a new internal
function called FetchToken().  This was done because the token table is now
searched up to three times: (1) to find the comment begin token; (2) to find
the comment end token; (3) to find the token after ther comment end token.
If the first token found is not a comment begin token, the token is returned
and no further searches are performed.

Prior to ParseToken(), ConvertEscAndMarkers() or ConvertParmMarkers(), which
process escape sequences and parameter markers, adds a space before and after
each end comment marker.  Otherwise, ParseToken(), which relies on spaces to
separater tokens, may not be able to correctly identify a comment begin or
coment end token.

------------------------------------------------------------------------