------------------------------------------------------------------------ r3319 | coomi01 | 2010-07-22 03:00:12 -0400 (Thu, 22 Jul 2010) | 11 lines
Review Board IP  http://reviewboard.ingres.prv/r/4814

Byte and VByte data should not be interpreted as
character strings, in particular where DByte char
sets are in use.

The change made is to copy the existing code removing
the CM macros for charset manipultaion in favour of
staright byte copying.

------------------------------------------------------------------------
r3320 | hanal04 | 2010-07-22 04:30:10 -0400 (Thu, 22 Jul 2010) | 12 lines

Cross integration of change 506168

Review Board IP  http://reviewboard.ingres.prv/r/4915

When a table name is specified on the command line optimizedb
looks in iitables and iisynonym for a reference to the supplied
table name. In a STAR DB iisysonym does not exist. This change
dupplicates the existing query (removing references to iisynonym)
if are connected to a Star DB.

------------------------------------------------------------------------