------------------------------------------------------------------------ r3320 | hanal04 | 2010-07-22 04:30:10 -0400 (Thu, 22 Jul 2010) | 12 lines
Cross integration of change 506168

Review Board IP  http://reviewboard.ingres.prv/r/4915

When a table name is specified on the command line optimizedb
looks in iitables and iisynonym for a reference to the supplied
table name. In a STAR DB iisysonym does not exist. This change
dupplicates the existing query (removing references to iisynonym)
if are connected to a Star DB.

------------------------------------------------------------------------
r3321 | schka24 | 2010-07-22 08:45:35 -0400 (Thu, 22 Jul 2010) | 24 lines

Add the create_compression DBMS config parameter, with values
NONE, DATA, or HIDATA.  The default is NONE.  This parameter
sets the compression to be applied for newly created tables
created with simple CREATE TABLE or DECLARE GLOBAL TEMPORARY
TABLE statements.  The parameter applies server-wide, to all sessions.

The create_compression parameter is ignored when creating tables
owned by $ingres (namely, catalogs).

The create_compression parameter does not apply to CREATE
TABLE AS SELECT.  CTAS already has several ways to control result
table compression.

There is no current way to apply default create-compression on a
session or table basis.  A WITH COMPRESSION clause for simple
CREATE is planned, but is not part of this IP. A new SET statement:
SET [no]CREATE_COMPRESSION
has been created to allow copydb/unloaddb to continue to use
their built-in assumption that create table creates uncompressed
tables.  This new statement is a bit of a dead end, since its job
can be done better by a WITH NOCOMPRESSION clause;  but until
such a clause exists, the SET statement is needed.

------------------------------------------------------------------------