------------------------------------------------------------------------ r3330 | schka01 | 2010-07-23 13:15:15 -0400 (Fri, 23 Jul 2010) | 13 lines
The fix (that I suggested) for b124007 causes HOQA to fail in
be/access/amxa02, leaving a dangling lock list in LK and causing
all sorts of downstream problems.

After further research, it appears that the real problem is
that LKconnect moves the "related lock list" from the original
LLB to the newly created SHARED LLB.  Related lock lists
don't belong on shared LLB's, they belong on handle LLB's
(the ones marked PARENT_SHARED).  Revert the LKshow fix,
and instead change LKconnect so that it leaves the related
lock list on the original LLB.

------------------------------------------------------------------------
r3331 | schka01 | 2010-07-23 13:15:30 -0400 (Fri, 23 Jul 2010) | 10 lines

The simple-aggregate processor dm2a.c will sometimes
create its own RCB for scanning a table, so that the
calling RCB position isn't adversely affected.  If there
are LOB's in the row, the LOB context pointer (BQCB)
has to be copied from the calling RCB into the new
one(s).  Or, if it can be guaranteed that the scan
cannot access the LOB(s) at all, we can mark the new
RCB's with "don't process LOB coupons".

------------------------------------------------------------------------