------------------------------------------------------------------------ r3332 | thaju02 | 2010-07-23 13:30:12 -0400 (Fri, 23 Jul 2010) | 15 lines
Cross integration of change 505331

Bug: 123887: ckpdb -d +j -l dbname after a second attempt fails
with E_DM9034, E_DM110B after applying patch 13749 and beyond.

Bug 123887 was introduced with incremental rolldb change 122216
which introduced write_jnlswitch() in dmfcpp.c to write out
a JNLSWITCH record to jnl before ckpdb updates ckp/jnl sequence
for offline checkpoint following archiver purge.

This change modifies write_jnlswitch(). If jnl file does not
exist checkpoint should continue.

------------------------------------------------------------------------
r3333 | schka01 | 2010-07-23 13:45:29 -0400 (Fri, 23 Jul 2010) | 18 lines

Outer join queries depend critically on the joinid to keep
predicates from the various ON clauses straight.  If a
predicate is constant-folded to TRUE or FALSE, we lose
the joinid, because there is no place to put a joinid in
a constant.

OPF expects join clause elements to be operators, so the
path of least resistance is to replace constant-folded
TRUE or FALSE with new built-in operators ii_true and
ii_false.  These operators take no arguments and simply
return true and false, but being operators they have a
place to put the joinid.

AND / OR / NOT constant folding in the parser now
recognizes ii_true and ii_false as constants in order to
take advantage of additional folding opportunities.

------------------------------------------------------------------------