------------------------------------------------------------------------ r3337 | horda03 | 2010-07-26 12:30:12 -0400 (Mon, 26 Jul 2010) | 45 lines
Review Board IP  http://reviewboard.ingres.prv/r/4997

Cross integration of change 506254

NB - Remove ^M from lines in Jamdefs.a64_win and
     Jamdefs.i64_win

The JAM build on Windows is very noisy due to the
shell commands executed by JAM always being echoed.

This is different behaviour from other platforms,
where "jam -dx" is used to instruct jam to display
the shell commands.

The reason why the shell commands are being displayed
on Windows ia because the variable JAMSHELL is
being defined (without the /Q qualifier), the shell
used to execute the OS comamnds always echo the
commands.

Rather than change the default behavior, added a
check on environment variable JAMSHELL_ECHO, if this
is not set to "OFF" then the JAMSHELL variable is
defined (just as it is prior to this change).

For greater flexibility, also allow JAMSHELL to
be defined in the user environment too.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------