------------------------------------------------------------------------ r3339 | schka01 | 2010-07-26 20:00:11 -0400 (Mon, 26 Jul 2010) | 9 lines
The ISAM row-duplicates testing needs the RCB extra
tuple buffer in several cases, but it was only allocating
it if the row is compressed.  Make sure it's allocated
for all cases that need the tuple buffer.

Checked btree and hash, they are OK (they allocate
it unconditionally when row-duplicates checking).

------------------------------------------------------------------------