------------------------------------------------------------------------ r3339 | schka01 | 2010-07-26 20:00:11 -0400 (Mon, 26 Jul 2010) | 9 lines
The ISAM row-duplicates testing needs the RCB extra
tuple buffer in several cases, but it was only allocating
it if the row is compressed.  Make sure it's allocated
for all cases that need the tuple buffer.

Checked btree and hash, they are OK (they allocate
it unconditionally when row-duplicates checking).

------------------------------------------------------------------------
r3340 | stephenb | 2010-07-26 23:10:20 -0400 (Mon, 26 Jul 2010) | 45 lines

Bug Number: 124107

rns.start
     124107
     (GENERIC)
      sometimes, if a dynamic cursor is defined to select from a sequence, and that sequence
      has been recently dropped and re-created, a fetch operation on the cursor may fail with
      either a sequencer SEGV or QEF errors
rns.end

Description of changes:
desc.start
The problem occurs because the cursor "FETCH" operation discovers during query execution
that it cannot find the iisequence record that matches the sequence ID in the query plan;
this causes QEF to return a "bad plan" error, which in turn causes the sequencer to re-try
the query, forcing OPF to generate a new plan. The sequencer does not
expect to do a "bad plan" re-try on a "FETCH" operation (because it assumes the plan
was refreshed on the OPEN CURSOR); this ultimately leads to it re-setting the query mode,
which causes it to go through the "parse" path on the re-try. This will
fail because "FETCH" operations do not contain query text. In the bug example it fails
with  SEGV because references are made to a non-existent query text pointer.

A "FETCH" operation should never fail with a bad query plan in the first place, because
the plan should be verified and locked during the "OPEN CURSOR"; so the cause of
the issue is that a stale query plan has not been correctly re-generated during the
"OPEN CURSOR". This occurs because the iisequence validation is usually done in the
parser, but during "OPEN CURSOR" the parser does not detect that a sequence is being
used. Cursors are a particularly devious means to get at a sequence and demonstrate
that the parser may not be the best place to validate a specific iisequence record;
since the query plan contains the sequence ID, we cannot reliably validate the record
until we have the plan in-hand. The correct place to validate the iisequence record
is during QEF plan validation.

This change modifies qeq_validate_qp to check for the existence of an
iisequence record that correctly qualifies the query plan (with the right sequence ID).
If the record doesn't exist, the plan is invalidated. With this code in place, the
plan is invalidated and correctly re-generated when processing the "CURSOR OPEN.

The change also fixes SCF so that a SEGV will not occur if there are other
code bugs that cause a plan re-try on a cursor fetch.

Related Service Desk Issues: 145696

Build Considerations:NONE
desc.end
------------------------------------------------------------------------