------------------------------------------------------------------------ r3340 | stephenb | 2010-07-26 23:10:20 -0400 (Mon, 26 Jul 2010) | 45 lines
Bug Number: 124107

rns.start
     124107
     (GENERIC)
      sometimes, if a dynamic cursor is defined to select from a sequence, and that sequence
      has been recently dropped and re-created, a fetch operation on the cursor may fail with
      either a sequencer SEGV or QEF errors
rns.end

Description of changes:
desc.start
The problem occurs because the cursor "FETCH" operation discovers during query execution
that it cannot find the iisequence record that matches the sequence ID in the query plan;
this causes QEF to return a "bad plan" error, which in turn causes the sequencer to re-try
the query, forcing OPF to generate a new plan. The sequencer does not
expect to do a "bad plan" re-try on a "FETCH" operation (because it assumes the plan
was refreshed on the OPEN CURSOR); this ultimately leads to it re-setting the query mode,
which causes it to go through the "parse" path on the re-try. This will
fail because "FETCH" operations do not contain query text. In the bug example it fails
with  SEGV because references are made to a non-existent query text pointer.

A "FETCH" operation should never fail with a bad query plan in the first place, because
the plan should be verified and locked during the "OPEN CURSOR"; so the cause of
the issue is that a stale query plan has not been correctly re-generated during the
"OPEN CURSOR". This occurs because the iisequence validation is usually done in the
parser, but during "OPEN CURSOR" the parser does not detect that a sequence is being
used. Cursors are a particularly devious means to get at a sequence and demonstrate
that the parser may not be the best place to validate a specific iisequence record;
since the query plan contains the sequence ID, we cannot reliably validate the record
until we have the plan in-hand. The correct place to validate the iisequence record
is during QEF plan validation.

This change modifies qeq_validate_qp to check for the existence of an
iisequence record that correctly qualifies the query plan (with the right sequence ID).
If the record doesn't exist, the plan is invalidated. With this code in place, the
plan is invalidated and correctly re-generated when processing the "CURSOR OPEN.

The change also fixes SCF so that a SEGV will not occur if there are other
code bugs that cause a plan re-try on a cursor fetch.

Related Service Desk Issues: 145696

Build Considerations:NONE
desc.end
------------------------------------------------------------------------
r3341 | horda03 | 2010-07-27 04:45:11 -0400 (Tue, 27 Jul 2010) | 52 lines

Review Board IP  http://reviewboard.ingres.prv/r/5006

Cross integration of change 506216

genxml utility is aborting when the database being unloaded
contains byte data.

The xf_encode_buffer detects when the result buffer is nearly
full, and then allocates a larger buffer. The contents of the
old buffer were copied to the new buffer prior to the memory for
the old buffer being released, but the "*obuf" wasn't updated
to the new buffer address.

The testcase for this issue was for Windows, so in compiling I've
also cleaned up the compiler warnings.

xf.qsh

Added prototype for get_table_nr used in xfgenxml.sc.

xfgenxml.sc

Two functions were initialising "dtable" (one a local copy, the
other a static for the compilation unit), so defined a new static
function init_encode_tables to initialise dtable and ctable.

Removed unused variables (ver and mem).

Added approriate castsi (ii_decimal and j). Changed the type of i to u_i4 to
match ilen.

The MEmove() function accepts u_i2's for the size of the src and destiantion
buffers, so split the MEmove into workable chunks.

And finally the fix, update *obuf with the address of the newBuf.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------