------------------------------------------------------------------------ r3344 | hweho01 | 2010-07-27 12:45:16 -0400 (Tue, 27 Jul 2010) | 6 lines
The HOQA be/raat problems are seen in several Unix platforms, such as SEGV
or SIGBUS, because there are mismatches between the RAAT table structures
and the table control blocks in dmtcb.h after the recent changes for long
name and column encryption supports.

------------------------------------------------------------------------
r3345 | hweho01 | 2010-07-27 13:00:10 -0400 (Tue, 27 Jul 2010) | 7 lines

When the table is dropped in HOQA be/raat02.sep test, error
E_DM005D_TABLE_ACCESS_CONFLICT is issued from dm2u_destroy(), because the
reference count is greater than one. To avoid the error, need to call
dmpe_query_end() to close the table that is opened to generate the logical
keys in dmpe_begin_dml() for the PUT processing.

------------------------------------------------------------------------