------------------------------------------------------------------------ r3350 | drivi01 | 2010-07-28 10:00:28 -0400 (Wed, 28 Jul 2010) | 7 lines
Ice feature is obsolete and no longer supported.

These changes removes/disables ICE in installer,
upgrades, embedded install, post installer,
build process and Visual DBA Tools.

------------------------------------------------------------------------
r3351 | schka01 | 2010-07-28 10:00:44 -0400 (Wed, 28 Jul 2010) | 16 lines

The 10.0 code for standard (data) compression uses a
"compression control" array.  This array is allocated before
many details of the table are known, so it's allocated based
on the number of attributes and whether the table might have
row-versioning or not.

The modify code was assuming that the modify result did not
need row-versioning.  This is true in a high level sense,
but as far as the compression control array is concerned,
there is still row-versioning until we rewrite the table
attributes with all the versioned atts dropped.  So,
make sure that modify (and create index) pass the current
relversion to the sizing routine, so that it will allocate
enough space for the control array.

------------------------------------------------------------------------