------------------------------------------------------------------------ r3367 | kiria01 | 2010-07-30 12:00:25 -0400 (Fri, 30 Jul 2010) | 43 lines
Scalar subquery cardinalty test too early.

With the move of aggregate functions including the SINGLETON
into derived tables, the SINGLETON aggregate is being finalized
prior to deselection of any rows that don't join.

Background Information:

Reported by customer on 10.0.0
Recreated here on head revs main

Documentation Impact:

None

The essence of this change is that the SINGLETON aggregate
responsible for raising the error no longer does so. Instead
it marks the potentially errant records and leaves the error
to be raised by the function SINGLECHK. The record is marked
by setting the ADF_SING_BIT in the NULL indicator byte of the
column passed to SINGLETON. This split allows the results of
the presence of a duplicated candidate to be represented.
The SINGLECHK then is only called after restrictions and any
joins have been applied.

pslssflat.c has the code added to introduce SINGLECHK into the
parse tree.

adeexecute.c has the code for SINGLECHK added and SINGLETON
changed as described above.

adiresty.c restores the NULLability of SINGLETON

adgoptab.roc, fi_defn.txt, adffiids.h, adfops.h changed to
add operator SINGLECHK.

pslsgram.yi, adeexecute.c, adugreatest.c, apidatav.c, adf.h
Are changed to tidy up the use of the null bit ADF_NVL_BIT
so that setting or testing it would bear deal with the bit and
not the byte unless so stated.

------------------------------------------------------------------------
r3368 | thogo01 | 2010-07-30 15:30:13 -0400 (Fri, 30 Jul 2010) | 5 lines

OpenAPI was not using the correct message buffer size when
a buffer is re-used off the free queue.  Use the size of the
buffer retrieved rather than using the minimum default size.

------------------------------------------------------------------------